// src/pages/cart/enum.ts
enum TagNames {
  DIV = 'div',
  H2 = 'h2',
  P = 'p',
  BUTTON = 'button',
  A = 'a',
  IMG = 'img',
}

enum Styles {
  CONTAINER = 'cart',
  TITLE = 'cart__title',
  LINE_ITEM_CONTAINER = 'cart__item-container',
  PROMO_GROUP = 'cart__promo-group',
  BTN_GROUP = 'cart__btn-group',
  BTN = 'cart__btn',
  BTN_ORDER = 'cart__btn--order',
  BTN_CLEAR = 'cart__btn--clear',
  ORDER_GROUP = 'cart__order-group',
  PRICE_WRAPPER = 'cart__price-wrapper',
  PRICE = 'cart__price',
  DISCOUNT_PRICE = 'cart__price-discount',
  PRICE_DISIBLED = 'cart__price--disabled',
}

enum Content {
  TITLE = 'Cart',
  BTN_ORDER = 'Make an order',
  BTN_CLEAR = 'Clear cart',
  BTN_CONTINUE = 'Continue shopping',
}

export { TagNames, Styles, Content };